fixed table view cell - title, time + body
