--- Journal.CoreData/Journal.CoreData/Entry+Convenience.swift
@@ -1,20 +1,0 @@
-//
-//  Entry+Convenience.swift
-//  Journal.CoreData
-//
-//  Created by beth on 2/24/20.
-//  Copyright © 2020 elizabeth wingate. All rights reserved.
-//
-
-import Foundation
-import CoreData
-
-extension Entry {
- convenience init(title: String, bodyText: String, timeStamp: Date = Date(), identifier: String = "UUID().uuidString", context: NSManagedObjectContext = CoreDataStack.shared.mainContext){
-    self.init(context: context)
-    self.title = title
-    self.bodyText = bodyText
-    self.timeStamp = timeStamp
-    self.identifier = identifier
-  }
-}
