+++ Journal.CoreData/Journal.CoreData/New Group/EntriesTableViewController.swift
