+++ Journal.CoreData/Journal.CoreData/New Group/EntryDetailViewController.swift
